add proxy classes

+++ concrete5.7.3.1/concrete/authentication/concrete/controller.php
@@ -21,10 +21,13 @@ class Controller extends AuthenticationTypeController
 
     public function deauthenticate(User $u)
     {
-        list($uID, $authType, $hash) = explode(':', $_COOKIE['ccmAuthUserHash']);
-        if ($authType == 'concrete') {
-            $db = Loader::db();
-            $db->execute('DELETE FROM authTypeConcreteCookieMap WHERE uID=? AND token=?', array($uID, $hash));
+        $cookie = array_get($_COOKIE, 'ccmAuthUserHash', '');
+        if ($cookie) {
+            list($uID, $authType, $hash) = explode(':', $cookie);
+            if ($authType == 'concrete') {
+                $db = Loader::db();
+                $db->execute('DELETE FROM authTypeConcreteCookieMap WHERE uID=? AND token=?', array($uID, $hash));
+            }
         }
     }
 
@@ -79,7 +82,9 @@ class Controller extends AuthenticationTypeController
     private function genString($a = 16)
     {
         if (function_exists('mcrypt_create_iv')) {
-            return bin2hex(mcrypt_create_iv($a, MCRYPT_DEV_URANDOM));
+            // Use /dev/urandom if available, otherwise fall back to PHP's rand.
+            // http://php.net/manual/en/function.mcrypt-create-iv.php#117047
+            return bin2hex(mcrypt_create_iv($a, MCRYPT_DEV_URANDOM|MCRYPT_RAND));
         } elseif (function_exists('openssl_random_pseudo_bytes')) {
             return bin2hex(openssl_random_pseudo_bytes($a));
         }
@@ -134,7 +139,7 @@ class Controller extends AuthenticationTypeController
                 //generate hash that'll be used to authenticate user, allowing them to change their password
                 $h = new \Concrete\Core\User\ValidationHash;
                 $uHash = $h->add($oUser->uID, intval(UVTYPE_CHANGE_PASSWORD), true);
-                $changePassURL = BASE_URL . View::url(
+                $changePassURL = View::url(
                         '/login',
                         'callback',
                         $this->getAuthenticationType()->getAuthenticationTypeHandle(),
@@ -143,13 +148,21 @@ class Controller extends AuthenticationTypeController
 
                 $mh->addParameter('changePassURL', $changePassURL);
 
-                if (defined('EMAIL_ADDRESS_FORGOT_PASSWORD')) {
-                    $mh->from(EMAIL_ADDRESS_FORGOT_PASSWORD, t('Forgot Password'));
-                } else {
+                $fromEmail = (string) Config::get('concrete.email.forgot_password.address');
+                if (!strpos($fromEmail, '@')) {
                     $adminUser = UserInfo::getByID(USER_SUPER_ID);
                     if (is_object($adminUser)) {
-                        $mh->from($adminUser->getUserEmail(), t('Forgot Password'));
+                        $fromEmail = $adminUser->getUserEmail();
+                    } else {
+                        $fromEmail = '';
                     }
+                }
+                if ($fromEmail) {
+                    $fromName = (string) Config::get('concrete.email.forgot_password.name');
+                    if ($fromName === '') {
+                        $fromName = t('Forgot Password');
+                    }
+                    $mh->from($fromEmail, $fromName);
                 }
 
                 $mh->addParameter('siteName', Config::get('concrete.site'));
@@ -236,11 +249,17 @@ class Controller extends AuthenticationTypeController
     {
         $post = $this->post();
 
-        if (!isset($post['uName']) || !isset($post['uPassword'])) {
+        if (empty($post['uName']) || empty($post['uPassword'])) {
             throw new Exception(t('Please provide both username and password.'));
         }
         $uName = $post['uName'];
         $uPassword = $post['uPassword'];
+
+        /** @type \Concrete\Core\Permission\IPService $ip_service */
+        $ip_service = \Core::make('ip');
+        if ($ip_service->isBanned()) {
+            throw new \Exception($ip_service->getErrorMessage());
+        }
 
         $user = new User($uName, $uPassword);
         if (!is_object($user) || !($user instanceof User) || $user->isError()) {
@@ -254,6 +273,12 @@ class Controller extends AuthenticationTypeController
                             'This account has not yet been validated. Please check the email associated with this account and follow the link it contains.'));
                     break;
                 case USER_INVALID:
+                    // Log failed auth
+                    $ip_service->logSignupRequest();
+                    if ($ip_service->signupRequestThreshholdReached()) {
+                        $ip_service->createIPBan();
+                        throw new \Exception($ip_service->getErrorMessage());
+                    }
                     if (Config::get('concrete.user.registration.email_registration')) {
                         throw new \Exception(t('Invalid email address or password.'));
                     } else {
